HHH-15415 Move the JaxbXmlFormatMapper as well for consistency
