+++ hibernate-core/src/main/java/org/hibernate/type/jaxb/JaxbXmlFormatMapper.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.type;
+package org.hibernate.type.jaxb;
 
 import java.io.StringReader;
 import java.io.StringWriter;
@@ -22,6 +22,7 @@ import javax.xml.namespace.QName;
 
 import org.hibernate.internal.util.ReflectHelper;
 import org.hibernate.internal.util.collections.CollectionHelper;
+import org.hibernate.type.FormatMapper;
 import org.hibernate.type.descriptor.WrapperOptions;
 import org.hibernate.type.descriptor.java.BasicPluralJavaType;
 import org.hibernate.type.descriptor.java.JavaType;
@@ -41,7 +42,7 @@ import org.w3c.dom.Node;
 /**
  * @author Christian Beikov
  */
-public class JaxbXmlFormatMapper implements FormatMapper {
+public final class JaxbXmlFormatMapper implements FormatMapper {
 
 	public static final String SHORT_NAME = "jaxb";
